Rename client and cleanup
