--- plasma_cash/decoded.py
@@ -1,1 +1,0 @@
-dec=b'\xf8Z\x80\x02\x94\x8c\x08\xe9\\q\x90kE\xbf\xfc\x9e\xc3\x8f\xb3\x9d3v`\x9b\xe9\xb8A\xb6\xaai\xe3\xa9nI<\x15~\xcaU,h\x1e\x1a\xe3J\xdf\xc3\x84\x8d\x9f\x9a&\xbf\x1a\xd1\x13\x99\x04\x005\x1e\xa5\xf5\xdd)\xab2\x8a\x14$\xea\x98\xa4\xe3\xba\xde5&\x9b\xe4\x97\xa3\xc07\xcc\xbe{{%\xc2w\x1b'
